camb

+++ RobotManager/src/com/gdc/nms/robot/util/DeleteServiceController.java
@@ -6,12 +6,16 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.logging.Logger;
+
+import javax.swing.JOptionPane;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.log4j.Logger;
 
 import com.gdc.nms.robot.gui.DeleteDirectory;
 import com.gdc.nms.robot.gui.RobotManager;
+import com.gdc.nms.robot.gui.util.process.JavaProcess;
+import com.gdc.nms.robot.gui.util.process.JavaProcessInfo;
 import com.gdc.nms.robot.util.indexer.AppInformation;
 import com.gdc.nms.robot.util.jade.InitPlataform;
 import com.gdc.robothelper.webservice.robot.CreatorRobotWebService;
@@ -23,7 +27,7 @@ import jade.core.AID;
 
 public class DeleteServiceController {
 	private AppInformation app;
-	private static final Logger LOGGER=Logger.getLogger(DeleteServiceController.class.toString());
+	private static final Logger LOGGER=Logger.getLogger(DeleteServiceController.class);
 
 	
 	public DeleteServiceController(AppInformation app){
@@ -32,18 +36,24 @@ public class DeleteServiceController {
 	
 	public boolean deleteService(){
 //			RobotManager.StopScanServices();
-		
-			long idRobot = app.getIdRobot();
-			LOGGER.info("Starting to Delete Robot id :"+idRobot);
-			boolean continueProcess=false;
-			continueProcess=CreatorRobotWebService.deleteRobot(idRobot);
-			LOGGER.info("robot webservices is deleted :"+continueProcess);
-			if(continueProcess){
-				continueProcess=deleteServices(app.getFolderPath());
-				LOGGER.info("moving folder to trash "+continueProcess);
-			}
+			LOGGER.info("try to stop robot for services "+app.getAppName());
+			if(stopRobot()){
+				long idRobot = app.getIdRobot();
+				LOGGER.info("Starting to Delete Robot id :"+idRobot);
+				boolean continueProcess=false;
+				continueProcess=CreatorRobotWebService.deleteRobot(idRobot);
+				LOGGER.info("robot webservices is deleted :"+continueProcess);
+				if(continueProcess){
+					continueProcess=deleteServices(app.getFolderPath());
+					LOGGER.info("moving folder to trash "+continueProcess);
+				}
 //			RobotManager.StartScanServices();
-			return continueProcess;
+				return continueProcess;
+				
+			}else{
+				JOptionPane.showMessageDialog(null,"No es posible detener el robot del servicio "+app.getAlias()+"es necesario detenerlo manualmente  y volver a intertarlo");
+				return false;
+			}
 	}
 	
 	
@@ -53,10 +63,38 @@ public class DeleteServiceController {
 		return false;
 	}
 	
+	
+	private boolean  stopRobot(){
+		if(app.isServicesRunning()){
+			System.out.println("services is running"+app.getAppName());
+			long pid = app.getPID();
+			if(JavaProcess.isValidJavaProceesId(pid)){
+				if(RobotManager.stopJar(pid)){
+					return true;
+				}
+			}else{
+				ArrayList<JavaProcessInfo> allJavaRobotProces = JavaProcess.getALLJavaRobotProces();
+				for (JavaProcessInfo javaProcessInfo : allJavaRobotProces) {
+					String commandLine = javaProcessInfo.getCommandLine();
+					if(commandLine.contains(app.getAppName()) ){
+						long processid = javaProcessInfo.getProcessid();
+						if(RobotManager.stopJar(processid)){
+							return true;
+						}
+					}
+				}
+			}
+			
+		}else{
+			return true;
+		}
+		return false;
+	}
+	
 	private boolean deleteServices(String serviceFolderPath){
 		Path serviceFolder = Paths.get(serviceFolderPath);
 		try {
-			return moveServicesToDelete(serviceFolder);
+			return deleteServicesFolder(serviceFolder);
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
@@ -64,26 +102,26 @@ public class DeleteServiceController {
 	}
 	
 	
-	private boolean moveServicesToDelete(Path serviceFolder) throws IOException{
-		Path servicesFolderPath = RobotManager.getServicesFolderPath();
-		createTrashFolder(servicesFolderPath);
-		Path trashFolder = servicesFolderPath.resolve("TRASH");
-		String name = serviceFolder.getFileName().toString();
-		String dateForFilesChanges = RobotManager.getDateForFilesChanges();
-		System.out.println(dateForFilesChanges);
-		Path resolve = serviceFolder.resolve("bot-1.0.jar");
-		ArrayList<String> validBotFiles = AppExaminator.getValidBotFiles(serviceFolder);
-		for (String string : validBotFiles) {
-			File botFile=new File(string);
-			Files.move(botFile.toPath(),resolve.resolveSibling("_D"+botFile.getName()));
-		}
-//		Path move = Files.move(resolve, resolve.resolveSibling("_Dvot-1.0.jar"));
-//		System.out.println("move"+move);
-		Path destinationPath = trashFolder.resolve(name+"_"+dateForFilesChanges);
-		if(Files.exists(destinationPath)){
-			Files.delete(destinationPath);
-		}
-		FileUtils.copyDirectoryToDirectory(serviceFolder.toFile(), destinationPath.toFile());
+	private boolean deleteServicesFolder(Path serviceFolder) throws IOException{
+//		Path servicesFolderPath = RobotManager.getServicesFolderPath();
+//		createTrashFolder(servicesFolderPath);
+//		Path trashFolder = servicesFolderPath.resolve("TRASH");
+//		String name = serviceFolder.getFileName().toString();
+//		String dateForFilesChanges = RobotManager.getDateForFilesChanges();
+//		System.out.println(dateForFilesChanges);
+//		Path resolve = serviceFolder.resolve("bot-1.0.jar");
+//		ArrayList<String> validBotFiles = AppExaminator.getValidBotFiles(serviceFolder);
+//		for (String string : validBotFiles) {
+//			File botFile=new File(string);
+//			Files.move(botFile.toPath(),resolve.resolveSibling("_D"+botFile.getName()));
+//		}
+////		Path move = Files.move(resolve, resolve.resolveSibling("_Dvot-1.0.jar"));
+////		System.out.println("move"+move);
+//		Path destinationPath = trashFolder.resolve(name+"_"+dateForFilesChanges);
+//		if(Files.exists(destinationPath)){
+//			Files.delete(destinationPath);
+//		}
+//		FileUtils.copyDirectoryToDirectory(serviceFolder.toFile(), destinationPath.toFile());
 		DeleteDirectory.delete(serviceFolder.toFile());
 		return true;
 	}
@@ -108,7 +146,7 @@ public class DeleteServiceController {
 		Path path = Paths.get("C:\\Users\\senrigan\\Documents\\pruebas\\GDC\\RobotScript\\data\\DYP");
 		boolean moveServicesToDelete;
 		try {
-			moveServicesToDelete = te.moveServicesToDelete(path);
+			moveServicesToDelete = te.deleteServicesFolder(path);
 			System.out.println(moveServicesToDelete);
 		} catch (IOException e) {
 			e.printStackTrace();
